+++ src/Quarry.Benchmarks/Benchmarks/CteProjectionBenchmarks.cs
@@ -33,7 +33,7 @@ public class CteProjectionBenchmarks : BenchmarkBase
             results.Add(new OrderIdTotalDto
             {
                 OrderId = reader.GetInt32(0),
-                Total = reader.GetDecimal(1)
+                Total = reader.GetDouble(1)
             });
         }
         return results;
@@ -86,7 +86,7 @@ public class CteProjectionBenchmarks : BenchmarkBase
             results.Add(new OrderIdTotalDto
             {
                 OrderId = reader.GetInt32(0),
-                Total = reader.GetDecimal(1)
+                Total = reader.GetDouble(1)
             });
         }
         return results;
